refactor(core): ♻️ move react-hook-form validateField tests to __tests__
