+++ packages/core/src/logic/validation/native-validation/index.ts
@@ -191,3 +191,5 @@ function createNativeValidatorSequencer(
   }
   return nativeValidatorSequencer
 }
+
+export default nativeValidateSingleField
